// src/utils/types.ts
export interface Current {
	time: string;
	interval: number;
	temperature_2m: number;
	relative_humidity_2m: number;
	apparent_temperature: number;
	is_day: number;
	precipitation: number;
	rain: number;
	showers: number;
	snowfall: number;
	weather_code: number;
	cloud_cover: number;
	pressure_msl: number;
	surface_pressure: number;
	wind_speed_10m: number;
	wind_direction_10m: number;
	wind_gusts_10m: number;
}